requirements.txt
asgiref==3.2.7
astroid==2.4.1
colorama==0.4.3
coverage==5.1
dj-database-url==0.5.0
Django==3.0.6
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.0
et-xmlfile==1.0.1
flake8==3.8.3
gunicorn==20.0.4
importlib-metadata==1.6.1
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.3
psycopg2==2.8.5
PuLP==2.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.12.1
zipp==3.1.0